Refactor papersolver into package
